+++ net/Buffer.h
@@ -22,7 +22,7 @@ public:
 	std::string retrieveAllAsString();
 	//how many data in the buffer now
 	int readbleBytes();
-	ssize_t readFd(int fd, int * savedErrno);
+	int readFd(int fd, int * savedErrno);
 private:
 	const char * begin() const
 	{
@@ -38,7 +38,7 @@ private:
 	{
 		return buffer_.capacity()-writeIndex_;
 	}
-	void ensureWritebale(ssize_t len);
+	void ensureWritebale(int len);
 
 	std::vector<char> buffer_;
 	int readIndex_;
